framework for items, basic storage
`Artemis-Collectibles` Scene, look at the first 2 cubes and hit interact/e to pick them up.

Items are created as scriptable objects, in the `Collectibles` folder, put in a list in `Collectible Item Registry` for possible use in saving/loading inventory items?

They're registered to specific Game Objects in the scene with the `PlacedCollectibleItem` component.

Debug logs the item name and description, removes the gameobject, and adds the item to the player "inventory"-- for now just a list of `InventoryCollectibleItem`