Finished the script for Pause Manu, but menu is weird. Talk to India...
The script itself does the job it's supposed but the UI menu itself is kind of compressed or stretched so it looks weird. If you press the buttons, they do the functions. But we need to fix that part of the UI. In other words, both scipts work fine, but fix the UI doc.