Catapult Launch Script
Added a Script that allows one to guide the trajectory of the launched object by inputting -

The Distance to be Travelled
The Height Difference Between the Initial and Final Position
and the Angle of the Launch.

The velocity necessary will be calculated by the script and then applied to the object. Currently, the script does not account for the catapult model itself, nor does it account for the direction of launch, assuming the movement will be done on the x-y plane.