Cleaner Catapult Code
Modified the CatapultLaunch, CatapultRotation, CatapultAmmo, CogTurning, and TouchGrass scripts so that the code is neater and more optimized.
Also added the AudioController & AnimationController scripts to reduce redundancy now and in the future.
Also also added an audio to the Rotation script so that the catapult makes a stone-grinding noise while turning. Will add more in the future.