Console Commands Update
Made some small changes to a couple of commands.

Teleport: Added some catches for user-error, such as if the user inputs an incorrect type, an incorrect amount of arguments, or some other invalid input. Also added the auto-completer for the pre-set locations.

IgnoreDialogue: Made it so that using ignoreDialogue while in the middle of dialogue will end the dialogue early. You still have to press enter after closing the console to completely end the convo. Side note: Also made it so you can't further the conversation while in the console.