Terrain textures and layers
Added textures for Rocky Jumpy soil and gravel, also lava with and without crusties. Crust lava should be used only if there's a large enough body of it to not have crust overlap the edges of the pool. I modified the base terrain layer and added another in Rocky Jumpy Road to use rocky jumpy soil and gravel.